+++ pkgs/http/lib/src/base_request.dart
@@ -133,13 +133,13 @@ abstract class BaseRequest {
       var response = await client.send(this);
       var stream = onDone(response.stream, client.close);
 
-      if (response is BaseResponseV2) {
+      if (response is BaseResponseWithUrl) {
         return StreamedResponseV2(ByteStream(stream), response.statusCode,
             contentLength: response.contentLength,
             request: response.request,
             headers: response.headers,
             isRedirect: response.isRedirect,
-            url: (response as BaseResponseV2).url,
+            url: (response as BaseResponseWithUrl).url,
             persistentConnection: response.persistentConnection,
             reasonPhrase: response.reasonPhrase);
       } else {