+++ pkgs/http/test/io/request_test.dart
@@ -46,8 +46,8 @@ void main() {
     final response = await request.send();
 
     expect(response.statusCode, equals(302));
-    expect(
-        (response as http.BaseResponseV2).url, serverUrl.resolve('/redirect'));
+    expect((response as http.BaseResponseWithUrl).url,
+        serverUrl.resolve('/redirect'));
   });
 
   test('with redirects', () async {
@@ -56,7 +56,7 @@ void main() {
     expect(response.statusCode, equals(200));
     final bytesString = await response.stream.bytesToString();
     expect(bytesString, parse(containsPair('path', '/')));
-    expect((response as http.BaseResponseV2).url, serverUrl.resolve('/'));
+    expect((response as http.BaseResponseWithUrl).url, serverUrl.resolve('/'));
   });
 
   test('exceeding max redirects', () async {